+++ app/src/main/java/com/example/jgjio_desktop/gostats/ViewableListAdapter.java
@@ -55,7 +55,7 @@ public class ViewableListAdapter extends PagedListAdapter<DataPoint, ViewableLis
         }
 
         void bindTo(int listIndex, DataPoint dataPoint) {
-            viewableDataPoint.setText(Double.toString(dataPoint.getValue()));
+            viewableDataPoint.setText(Double.toString(dataPoint.getValue().doubleValue()));
             positionDataPoint.setText(Integer.toString(listIndex + 1)); //statistical lists start at 1 instead of 0
 
             if (!dataPoint.isEnabled()) {
